+++ LogicBuilder.App.Bsl.Utils.Tests/PersistenceOperationsTest.cs
@@ -56,7 +56,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.AddChange(schoolRepository, student);
+            PersistenceOperationsUtils<StudentModel, Student>.AddChange(schoolRepository, student);
             await schoolRepository.SaveChangesAsync();
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
@@ -85,7 +85,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.AddChanges(schoolRepository, [student]);
+            PersistenceOperationsUtils<StudentModel, Student>.AddChanges(schoolRepository, [student]);
             await schoolRepository.SaveChangesAsync();
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
@@ -114,7 +114,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.AddGraphChange(schoolRepository, student);
+            PersistenceOperationsUtils<StudentModel, Student>.AddGraphChange(schoolRepository, student);
             await schoolRepository.SaveChangesAsync();
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
@@ -143,7 +143,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.AddGraphChanges(schoolRepository, [student]);
+            PersistenceOperationsUtils<StudentModel, Student>.AddGraphChanges(schoolRepository, [student]);
             await schoolRepository.SaveChangesAsync();
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
@@ -161,7 +161,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             var before = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == fullName)).SingleOrDefault();
 
             //act
-            PersistenceOperations<StudentModel, Student>.Delete(schoolRepository, s => s.FullName == fullName);
+            PersistenceOperationsUtils<StudentModel, Student>.Delete(schoolRepository, s => s.FullName == fullName);
             var after = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == fullName)).SingleOrDefault();
 
             //assert
@@ -189,7 +189,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.Save(schoolRepository, student);
+            PersistenceOperationsUtils<StudentModel, Student>.Save(schoolRepository, student);
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
             //assert
@@ -217,7 +217,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.Save(schoolRepository, [student]);
+            PersistenceOperationsUtils<StudentModel, Student>.Save(schoolRepository, [student]);
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
             //assert
@@ -245,7 +245,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.SaveGraph(schoolRepository, student);
+            PersistenceOperationsUtils<StudentModel, Student>.SaveGraph(schoolRepository, student);
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
             //assert
@@ -273,7 +273,7 @@ namespace LogicBuilder.App.Bsl.Utils.Tests
             };
 
             //act
-            PersistenceOperations<StudentModel, Student>.SaveGraphs(schoolRepository, [student]);
+            PersistenceOperationsUtils<StudentModel, Student>.SaveGraphs(schoolRepository, [student]);
             var result = (await schoolRepository.GetAsync<StudentModel, Student>(s => s.FullName == "Roger Milla", null, new SelectExpandDefinition([], [new SelectExpandItem("Enrollments")]))).Single();
 
             //assert